Feature: moved ErrorBoundary component to its own file
Made use of the ErrorBoundary in more places - catch errors without breaking everything
Added fallback screen to main error boundary
Cleaned up imports - always use component index
