+++ src/components/ErrorMessage/styles.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  text-align: center;
+`;
