+++ src/components/GameCard/index.tsx
@@ -5,8 +5,9 @@ import { Link } from 'react-router-dom';
 import { CANCEL_GAME, CancelGameMutation } from '../../graphql/mutations';
 import { formatEnum, getLoggedInUserId } from '../../utils';
 import Avatar from '../Avatar';
+import ErrorMessage from '../ErrorMessage';
 import { H3 } from '../HeaderText';
-import { Board, Button, Card } from '../index';
+import { Board, Button, Card, ErrorBoundary } from '../index';
 import { CardInner, ColorText, PlayerDiv, TeamDiv, TeamsContainer } from './styles';
 
 interface GameCardProps extends Game {
@@ -38,65 +39,68 @@ export default class GameCard extends React.Component<GameCardProps> {
   render() {
     const {id, redPlayer, yellowPlayer, winner, status, grid, style, createdAt, nextPlayer} = this.props;
     return (
-      <Card
-        style={style}
-        contentPadding={true}
-        title={formatEnum(status)}
-        date={createdAt}
-      >
-        <CardInner>
-          <TeamsContainer>
-            <TeamDiv>
-              <Link to={'/dashboard/' + id}>
-                <PlayerDiv>
-                  <Avatar
-                    activeColor={status === 'IN_PROGRESS' && nextPlayer === 'RED' ? 'RED' : undefined}
-                    size={50}
-                    imageUrl={
-                      redPlayer
-                        ? redPlayer.avatarUrl
-                        : null
-                    }
-                  />
-                  {GameCard.renderPlayerName(redPlayer)}
-                </PlayerDiv>
-              </Link>
-              <ColorText>Red {winner === 'RED' && <span>🏆</span>}</ColorText>
-            </TeamDiv>
-            <H3>VS</H3>
-            <TeamDiv>
-              <Link to={'/dashboard/' + id}>
-                <PlayerDiv>
-                  <Avatar
-                    activeColor={status === 'IN_PROGRESS' && nextPlayer === 'YELLOW' ? 'YELLOW' : undefined}
-                    size={50}
-                    imageUrl={
-                      yellowPlayer
-                        ? yellowPlayer.avatarUrl
-                        : null
-                    }
-                  />
-                  {GameCard.renderPlayerName(yellowPlayer)}
-                </PlayerDiv>
-              </Link>
-              <ColorText>Yellow {winner === 'YELLOW' && <span>🏆</span>}</ColorText>
-            </TeamDiv>
-          </TeamsContainer>
+      // TODO change message to message in card
+      <ErrorBoundary FallbackComponent={ErrorMessage}>
+        <Card
+          style={style}
+          contentPadding={true}
+          title={formatEnum(status)}
+          date={createdAt}
+        >
+          <CardInner>
+            <TeamsContainer>
+              <TeamDiv>
+                <Link to={'/dashboard/' + id}>
+                  <PlayerDiv>
+                    <Avatar
+                      activeColor={status === 'IN_PROGRESS' && nextPlayer === 'RED' ? 'RED' : undefined}
+                      size={50}
+                      imageUrl={
+                        redPlayer
+                          ? redPlayer.avatarUrl
+                          : null
+                      }
+                    />
+                    {GameCard.renderPlayerName(redPlayer)}
+                  </PlayerDiv>
+                </Link>
+                <ColorText>Red {winner === 'RED' && <span>🏆</span>}</ColorText>
+              </TeamDiv>
+              <H3>VS</H3>
+              <TeamDiv>
+                <Link to={'/dashboard/' + id}>
+                  <PlayerDiv>
+                    <Avatar
+                      activeColor={status === 'IN_PROGRESS' && nextPlayer === 'YELLOW' ? 'YELLOW' : undefined}
+                      size={50}
+                      imageUrl={
+                        yellowPlayer
+                          ? yellowPlayer.avatarUrl
+                          : null
+                      }
+                    />
+                    {GameCard.renderPlayerName(yellowPlayer)}
+                  </PlayerDiv>
+                </Link>
+                <ColorText>Yellow {winner === 'YELLOW' && <span>🏆</span>}</ColorText>
+              </TeamDiv>
+            </TeamsContainer>
 
-          <Board gameId={id} grid={grid} isActive={GameCard.isGamePlayable(status, redPlayer, yellowPlayer)} />
+            <Board gameId={id} grid={grid} isActive={GameCard.isGamePlayable(status, redPlayer, yellowPlayer)} />
 
-          {status === 'NOT_STARTED' &&
-          <CancelGameMutation mutation={CANCEL_GAME}>
-            {(cancelGame: Function) => (
-              <Button
-                noMargin={true}
-                onClick={() => cancelGame({variables: {gameId: id}})}
-                text='Cancel Game'
-              />
-            )}
-          </CancelGameMutation>}
-        </CardInner>
-      </Card>
+            {status === 'NOT_STARTED' &&
+            <CancelGameMutation mutation={CANCEL_GAME}>
+              {(cancelGame: Function) => (
+                <Button
+                  noMargin={true}
+                  onClick={() => cancelGame({variables: {gameId: id}})}
+                  text='Cancel Game'
+                />
+              )}
+            </CancelGameMutation>}
+          </CardInner>
+        </Card>
+      </ErrorBoundary>
     );
   }
 }
