+++ src/components/index.ts
@@ -1,9 +1,13 @@
 export { default as Avatar } from './Avatar';
 export { default as Board } from './Board';
+export { default as Button } from './Button';
 export { default as Card } from './Card';
 export { default as Cell } from './Cell';
-export { default as Games } from './Games';
-export { default as Button } from './Button';
 export { default as ContainerWrapper } from './ContainerWrapper';
+export { default as ErrorMessage } from './ErrorMessage';
+export { default as ErrorScreen } from './ErrorScreen';
+export { default as GameCard } from './GameCard';
+export { default as Games } from './Games';
 export { default as Header } from './Header';
+export { ErrorBoundary } from './ErrorBoundary';
 export { H1, H2, H3, H4 } from './HeaderText';
