+++ src/containers/Login/index.tsx
@@ -1,10 +1,8 @@
 // Login Page
 
 import * as React from 'react';
+import { Button, ContainerWrapper, H1 } from '../../components';
 import { SLACK_CLIENT_URL } from '../../constants';
-import { ContainerWrapper } from '../../components';
-import { H1 } from '../../components/HeaderText';
-import { default as Button } from '../../components/Button';
 
 const Login = () => {
   return (
@@ -13,7 +11,7 @@ const Login = () => {
       <H1>NN</H1>
       <H1>EC</H1>
       <H1>T4</H1>
-      <Button text='Login with Slack' onClick={() => window.open(SLACK_CLIENT_URL, '_self')}/>
+      <Button text='Login with Slack' onClick={() => window.open(SLACK_CLIENT_URL, '_self')} />
     </ContainerWrapper>
   );
 };
